+++ common/src/main/java/de/mrjulsen/wires/WirePoints.java
@@ -1,5 +1,6 @@
 package de.mrjulsen.wires;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
-public record WirePoints(Vector3f[] vertices) { }
+public record WirePoints(Vector3d[] vertices) { }
