+++ common/src/main/java/de/mrjulsen/wires/decoration/WireDecorationRenderData.java
@@ -1,5 +1,5 @@
 package de.mrjulsen.wires.decoration;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
-public record WireDecorationRenderData(Vector3f worldPos, Vector3f direction, WireDecorationData data) {}
+public record WireDecorationRenderData(Vector3d worldPos, Vector3d direction, WireDecorationData data) {}
